v2: extract FE 2e assembly into reusable free functions; route NAORadialBasis through them
DRY follow-on to PR #81 (per design feedback). The
NAORadialBasis::assemble_J_fe_ / assemble_K_fe_ helpers were pure
FE-level assembly that did not need NAO context. They are promoted to
free functions in helfem::atomic::basis, in a new header
libhelfem/include/CoulombExchangeFE.h:

  arma::mat assemble_J_FE_one_multipole(
      const FEMRadialBasis & radial, int L, const arma::mat & P_FE);
  arma::mat assemble_K_FE_one_multipole(
      const FEMRadialBasis & radial, int L, const arma::mat & P_FE);
  arma::mat assemble_J_FE_one_multipole_yukawa(
      const FEMRadialBasis & radial, int L, double lambda,
      const arma::mat & P_FE);
  arma::mat assemble_K_FE_one_multipole_yukawa(
      const FEMRadialBasis & radial, int L, double lambda,
      const arma::mat & P_FE);

Same semantics as before: bare radial operator at a single multipole
L (no 4*pi/(2L+1) prefactor, no Gaunt; sign convention positive bare
integral). Header-only, ~180 LOC, takes the same inputs the
NAORadialBasis methods used to take internally.

NAORadialBasis::coulomb_radial / exchange_radial / coulomb_radial_yukawa
/ exchange_radial_yukawa now collapse to one-line wrappers that
C-transform around the free function:

  arma::mat coulomb_radial(int k, const NAORadialBasis & other,
                            const arma::mat & D) const {
    ...
    return C_.t() * assemble_J_FE_one_multipole(fem, k, P_FE) * C_;
  }

Net diff in NAORadialBasis.h: -113 LOC (the assemble_J_fe_ /
assemble_K_fe_ / r_small_ / r_big_ / in_element_tei_ / exchange_tei_
private helpers all go away, replaced by the free-function call). The
exchange_tei forward-declaration that was in NAORadialBasis.h also
moves to CoulombExchangeFE.h.

Validation: ran the same end-to-end test suite from PR #81
(8-LIP, 10 exponential elements, Z=2 He at l=0):

  [Test 1: J at L=0 vs 4*pi * nao.coulomb_radial]
    ||J_sadatom - 4pi*J_nao||_inf  = 1.776e-15  (same as before)
    rel err                         = 1.497e-16  (same)
  [Test 1b: K at L=0 vs -nao.exchange_radial]
    ||K_sadatom + K_nao||_inf      = 5.551e-17  (same)
  [Test 2: C-transform invariance]
    ||J_small - C^T J_FE C||_inf   = 0.000e+00  (same)
    ||K_small - C^T K_FE C||_inf   = 0.000e+00  (same)
  [Test 3: L=2 (higher multipole)]
    ||J_L2 - J_L2^T||_inf          = 6.939e-18  (same)
    ||K_L2 - K_L2^T||_inf          = 6.939e-18  (same)
  [Test 4: cross-channel J (Csmall, Cb), L=0]
    ||J_cross - C^T J_FE(C_b D C_b^T) C||_inf = 0.000e+00  (same)
  [Test 5: Yukawa Coulomb at lambda=0.5 runs]
    ||J_yuk||_fro                   = 1.841855e+00  (same)

Bit-for-bit identical to the pre-refactor results -- pure code
motion, no semantic change.

Verified: full build clean; gaunt_test / sphtest / legendre_test
pass; NAO export example (PR #55) matches prior numerics bit-for-bit
(He HF = -2.86167999561 Eh, err 3.59e-12).

What this PR does NOT do (deferred): the sadatom and atomic
TwoDBasis::coulomb / TwoDBasis::exchange routines still carry their
own inline per-multipole assembly (with the precomputed disjoint /
prim_tei caches for SCF-loop performance). Migrating those onto the
new free functions requires either (a) preserving the caching via a
cache-aware overload, or (b) accepting the cache-free overhead --
both are doable but are a separate refactor with SCF regression
implications. Next PR.

Also paves the way for the Cholesky migration (PR #82's
twoe_integral_cholesky): once the free functions accept the in-element
tensor in factored form, the unified contraction routine
(Sum_p L_p . P . L_p^T) replaces the direct n^4 path in both
NAORadialBasis and TwoDBasis.
